[refactor] src path 수정
